Cleanup Yolov8s and Yolov8s_world demo_utils
